Backend done

+++ project/src/main/java/bll/crud/BusBll.java
@@ -1,9 +1,11 @@
 package bll.crud;
 
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
+import bll.dtos.BusDTO;
+import bll.dtos.converters.BusConverter;
 import dal.entities.Bus;
 import dal.repositories.BusRepository;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
 
 import java.util.List;
 
@@ -11,35 +13,43 @@ import java.util.List;
 public class BusBll {
     @Autowired
     BusRepository busRepository;
+    @Autowired
+    private BusConverter converter;// = new BusConverter();
 
-    public List<Bus> getAllBuses() {
-        return busRepository.findAll();
+    public List<BusDTO> getAllBuses() {
+        return converter.entityListToDtoList(busRepository.findAll());
     }
 
-    public Bus getBusById(int id) {
+    public BusDTO getBusById(int id) {
         if (busRepository.findById(id).isPresent())
-            return busRepository.findById(id).get();
-        else
-            return null;
+            return converter.entityToDto(busRepository.findById(id).get());
+        return null;
     }
 
-    public void addBus(Bus bus) {
-        busRepository.save(bus);
+    public String addBus(BusDTO bus) {
+        busRepository.save(converter.dtoToEntity(bus));
+        return "BUS ADD SUCCESSFUL";
     }
 
-    public String updateBus(Bus bus) {
+    public String updateBus(BusDTO bus) {
         Bus updatedBus;
+        String reason = "Bus";
         if (busRepository.findById(bus.getId()).isPresent()) {
             updatedBus = busRepository.findById(bus.getId()).get();
             updatedBus.setLine(bus.getLine());
             updatedBus.setRoute(bus.getRoute());
             busRepository.save(updatedBus);
             return "BUS UPDATE SUCCESSFUL";
-        } else
-            return "BUS UPDATE FAILED";
+        }
+        return "BUS UPDATE FAILED: " + reason + " with this ID doesn't exist";
     }
 
-    public void deleteBus(int busId) {
-        busRepository.deleteById(busId);
+    public String deleteBus(int busId) {
+        String reason = "Bus";
+        if (busRepository.findById(busId).isPresent()) {
+            busRepository.deleteById(busId);
+            return "BUS DELETE SUCCESSFUL";
+    }
+        return "BUS DELETE FAILED: " + reason + " with this ID doesn't exist";
     }
 }
